Removing parent directories and refactoring
