+++ animals/cobra.py
