+++ animals/coralsnake.py
