+++ animals/cow.py
