+++ animals/diamondback.py
