+++ animals/gardensnake.py
